separated validation from index
